refactor: create handler package
