+++ src/handler/discord.js/commands/ChoiceCommandHandler.ts
@@ -1,7 +1,7 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
 import { ChoiceContent } from "@/src/entities/vo/ChoiceContent";
+import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
 import type { IUtilityLogic } from "@/src/logics/Interfaces/logics/IUtilityLogic";
-import type { SlashCommandHandler } from "@/src/routes/discordjs/handler/commands/SlashCommandHandler";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
