+++ src/handler/discord.js/commands/DiceCommandHandler.ts
@@ -1,24 +1,24 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
-import { ParrotMessage } from "@/src/entities/vo/ParrotMessage";
+import { DiceSides } from "@/src/entities/vo/DiceSides";
+import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
 import type { IUtilityLogic } from "@/src/logics/Interfaces/logics/IUtilityLogic";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
-import type { SlashCommandHandler } from "src/routes/discordjs/handler/commands/SlashCommandHandler";
 
 @injectable()
-export class ParrotCommandHandler implements SlashCommandHandler {
+export class DiceCommandHandler implements SlashCommandHandler {
 	@inject(LogicTypes.UtilityLogic)
 	private utilLogic!: IUtilityLogic;
 	isHandle(commandName: string): boolean {
-		return commandName === "parrot";
+		return commandName === "dice";
 	}
 
 	async handle(
 		interaction: ChatInputCommandInteraction<CacheType>,
 	): Promise<void> {
 		await interaction.reply(
-			await this.utilLogic.parrot(
-				new ParrotMessage(interaction.options?.getString("message", true)),
+			await this.utilLogic.dice(
+				new DiceSides(interaction.options?.getInteger("parameter", true)),
 			),
 		);
 	}
