+++ src/handler/discord.js/commands/HelpCommandHandler.ts
@@ -1,24 +1,23 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
-import { DiceSides } from "@/src/entities/vo/DiceSides";
+import { HelpCategory } from "@/src/entities/vo/HelpCategory";
+import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
 import type { IUtilityLogic } from "@/src/logics/Interfaces/logics/IUtilityLogic";
-import type { SlashCommandHandler } from "@/src/routes/discordjs/handler/commands/SlashCommandHandler";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
 
 @injectable()
-export class DiceCommandHandler implements SlashCommandHandler {
+export class HelpCommandHandler implements SlashCommandHandler {
 	@inject(LogicTypes.UtilityLogic)
 	private utilLogic!: IUtilityLogic;
 	isHandle(commandName: string): boolean {
-		return commandName === "dice";
+		return commandName === "help";
 	}
-
 	async handle(
 		interaction: ChatInputCommandInteraction<CacheType>,
 	): Promise<void> {
 		await interaction.reply(
-			await this.utilLogic.dice(
-				new DiceSides(interaction.options?.getInteger("parameter", true)),
+			await this.utilLogic.help(
+				new HelpCategory(interaction.options?.getString("category", true)),
 			),
 		);
 	}
