+++ src/handler/discord.js/commands/PointItemCommandHandler.ts
@@ -1,24 +1,24 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
 import { DiscordUserId } from "@/src/entities/vo/DiscordUserId";
+import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
 import type { IPointLogic } from "@/src/logics/Interfaces/logics/IPointLogic";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
-import type { SlashCommandHandler } from "src/routes/discordjs/handler/commands/SlashCommandHandler";
 
 @injectable()
-export class PointDrawCommandHandler implements SlashCommandHandler {
+export class PointItemCommandHandler implements SlashCommandHandler {
 	@inject(LogicTypes.PointLogic)
 	private pointLogic!: IPointLogic;
 
 	isHandle(commandName: string): boolean {
-		return commandName === "pointdraw";
+		return commandName === "pointitem";
 	}
 
 	async handle(
 		interaction: ChatInputCommandInteraction<CacheType>,
 	): Promise<void> {
 		await interaction.reply(
-			await this.pointLogic.drawItem(new DiscordUserId(interaction.user.id)),
+			await this.pointLogic.getItems(new DiscordUserId(interaction.user.id)),
 		);
 	}
 }
