+++ src/handler/discord.js/commands/ReminderListCommandHandler.ts
@@ -1,24 +1,24 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
 import { DiscordUserId } from "@/src/entities/vo/DiscordUserId";
-import type { IPointLogic } from "@/src/logics/Interfaces/logics/IPointLogic";
+import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
+import type { IReminderLogic } from "@/src/logics/Interfaces/logics/IReminderLogic";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
-import type { SlashCommandHandler } from "src/routes/discordjs/handler/commands/SlashCommandHandler";
 
 @injectable()
-export class PointItemCommandHandler implements SlashCommandHandler {
-	@inject(LogicTypes.PointLogic)
-	private pointLogic!: IPointLogic;
+export class ReminderListCommandHandler implements SlashCommandHandler {
+	@inject(LogicTypes.ReminderLogic)
+	private reminderLogic!: IReminderLogic;
 
 	isHandle(commandName: string): boolean {
-		return commandName === "pointitem";
+		return commandName === "reminderlist";
 	}
 
 	async handle(
 		interaction: ChatInputCommandInteraction<CacheType>,
 	): Promise<void> {
 		await interaction.reply(
-			await this.pointLogic.getItems(new DiscordUserId(interaction.user.id)),
+			await this.reminderLogic.list(new DiscordUserId(interaction.user.id)),
 		);
 	}
 }
