+++ src/handler/discord.js/commands/ReviewGachaCommandHandler.ts
@@ -1,17 +1,18 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
 import { DiscordUserId } from "@/src/entities/vo/DiscordUserId";
+import { GithubPullRequestId } from "@/src/entities/vo/GithubPullRequestId";
+import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
 import type { IPullRequestLogic } from "@/src/logics/Interfaces/logics/IPullRequestLogic";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
-import type { SlashCommandHandler } from "src/routes/discordjs/handler/commands/SlashCommandHandler";
 
 @injectable()
-export class ReviewListCommandHandler implements SlashCommandHandler {
+export class ReviewGachaCommandHandler implements SlashCommandHandler {
 	@inject(LogicTypes.PullRequestLogic)
 	private pullRequestLogic!: IPullRequestLogic;
 
 	isHandle(commandName: string): boolean {
-		return commandName === "reviewlist";
+		return commandName === "reviewgacha";
 	}
 
 	async handle(
@@ -19,7 +20,8 @@ export class ReviewListCommandHandler implements SlashCommandHandler {
 	): Promise<void> {
 		await interaction.deferReply();
 		await interaction.editReply(
-			await this.pullRequestLogic.findAssignedPullRequest(
+			await this.pullRequestLogic.randomAssign(
+				new GithubPullRequestId(interaction.options?.getInteger("id", true)),
 				new DiscordUserId(interaction.user.id),
 			),
 		);
