+++ src/handler/discord.js/commands/ReviewListCommandHandler.ts
@@ -1,18 +1,17 @@
 import { LogicTypes } from "@/src/entities/constants/DIContainerTypes";
 import { DiscordUserId } from "@/src/entities/vo/DiscordUserId";
-import { GithubPullRequestId } from "@/src/entities/vo/GithubPullRequestId";
+import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
 import type { IPullRequestLogic } from "@/src/logics/Interfaces/logics/IPullRequestLogic";
 import type { CacheType, ChatInputCommandInteraction } from "discord.js";
 import { inject, injectable } from "inversify";
-import type { SlashCommandHandler } from "src/routes/discordjs/handler/commands/SlashCommandHandler";
 
 @injectable()
-export class ReviewGachaCommandHandler implements SlashCommandHandler {
+export class ReviewListCommandHandler implements SlashCommandHandler {
 	@inject(LogicTypes.PullRequestLogic)
 	private pullRequestLogic!: IPullRequestLogic;
 
 	isHandle(commandName: string): boolean {
-		return commandName === "reviewgacha";
+		return commandName === "reviewlist";
 	}
 
 	async handle(
@@ -20,8 +19,7 @@ export class ReviewGachaCommandHandler implements SlashCommandHandler {
 	): Promise<void> {
 		await interaction.deferReply();
 		await interaction.editReply(
-			await this.pullRequestLogic.randomAssign(
-				new GithubPullRequestId(interaction.options?.getInteger("id", true)),
+			await this.pullRequestLogic.findAssignedPullRequest(
 				new DiscordUserId(interaction.user.id),
 			),
 		);
