+++ src/routes/discordjs/events/MessageReplyRouter.ts
@@ -1,8 +1,8 @@
 import { HandlerTypes } from "@/src/entities/constants/DIContainerTypes";
+import type { DiscordEventHandler } from "@/src/handler/discord.js/events/DiscordEventHandler";
 import type { DiscordEventRouter } from "@/src/routes/discordjs/events/DiscordEventRouter";
 import type { Client, Message } from "discord.js";
 import { injectable, multiInject } from "inversify";
-import type { DiscordEventHandler } from "src/routes/discordjs/handler/events/DiscordEventHandler";
 
 @injectable()
 export class MessageReplyRouter implements DiscordEventRouter {
