+++ src/routes/discordjs/events/SlashCommandRouter.ts
@@ -1,7 +1,7 @@
 import { InternalErrorMessage } from "@/src/entities/DiscordErrorMessages";
 import { HandlerTypes } from "@/src/entities/constants/DIContainerTypes";
+import type { SlashCommandHandler } from "@/src/handler/discord.js/commands/SlashCommandHandler";
 import type { DiscordEventRouter } from "@/src/routes/discordjs/events/DiscordEventRouter";
-import type { SlashCommandHandler } from "@/src/routes/discordjs/handler/commands/SlashCommandHandler";
 import type { Client } from "discord.js";
 import { injectable, multiInject } from "inversify";
