--- src/routes/discordjs/handler/commands/index.ts
@@ -1,18 +1,0 @@
-export { ReminderSetCommandHandler } from "./ReminderSetCommandHandler";
-export { ReminderListCommandHandler } from "./ReminderSetCommandHandler";
-export { ReminderDeleteCommandHandler } from "./ReminderSetCommandHandler";
-export { ReviewGachaCommandHandler } from "./ReviewGachaCommandHandler";
-export { ReviewListCommandHandler } from "./ReviewListCommandHandler";
-export { TranslateCommandHandler } from "./TranslateCommandHandlers";
-export { ChoiceCommandHandler } from "./ChoiceCommandHandler";
-export { DiceCommandHandler } from "./DiceCommandHandler";
-export { HelpCommandHandler } from "./HelpCommandHandler";
-export { MinecraftStartCommandHandler } from "./MinecraftStartCommandHandler";
-export { MinecraftStopCommandHandler } from "./MinecraftStartCommandHandler";
-export { ParrotCommandHandler } from "./ParrotCommandHandler";
-export { PointCheckCommandHandler } from "./PointCheckCommandHandler";
-export { PointExchangeCommandHandler } from "./PointCheckCommandHandler";
-export { PointDrawCommandHandler } from "./PointDrawCommandHandler";
-export { PointItemCommandHandler } from "./PointItemCommandHandler";
-export { TalkCommandHandler } from "./TalkCommandHandler";
-export { WaiwaiCommandHandler } from "./WaiwaiCommandHandler";
